Move the ejson wrapper to the list of public packages
Necessary because the values package (and all others which
use it) require the Settings type as parameter.
