+++ pkg/values/directory_test.go
@@ -21,7 +21,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/bedag/kusible/internal/wrapper/ejson"
+	"github.com/bedag/kusible/pkg/wrapper/ejson"
 	log "github.com/sirupsen/logrus"
 	"gotest.tools/assert"
 	"sigs.k8s.io/yaml"
