+++ pkg/values/values_test.go
@@ -20,7 +20,7 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/bedag/kusible/internal/wrapper/ejson"
+	"github.com/bedag/kusible/pkg/wrapper/ejson"
 	"gotest.tools/assert"
 	"sigs.k8s.io/yaml"
 )
